requirements.txt
Flask==3.1.1
tensorflow-cpu==2.19.0
opencv-python-headless==4.11.0.86  
mediapipe==0.10.21
numpy==1.26.4
keras==3.10.0
gunicorn==23.0.0  
